+++ api/v1alpha1/applyconfiguration/utils.go
@@ -268,6 +268,8 @@ func ForKind(kind schema.GroupVersionKind) interface{} {
 		return &apiv1alpha1.SMTPConfigApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("TemplateReference"):
 		return &apiv1alpha1.TemplateReferenceApplyConfiguration{}
+	case v1alpha1.SchemeGroupVersion.WithKind("TerminalRecordingRef"):
+		return &apiv1alpha1.TerminalRecordingRefApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("TerminalSharingConfig"):
 		return &apiv1alpha1.TerminalSharingConfigApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("TerminalSharingStatus"):